% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{Bin}
\alias{Bin}
\title{Bin numeric & factor vectors and data.frames}
\usage{
Bin(x, y, w, seg, name = "NONE", ...)
}
\arguments{
\item{x}{Object to be binned. See details.}

\item{y}{Dependent variable used for discretization. Only binary outcomes are
currently supported. The discretization algorithm maximizes information value
through binary partitions.}

\item{w}{An optional numeric weight variable.}

\item{seg}{An optional factor to be used for creating a segmented classing
object. See details.}

\item{...}{Optional arguments passed on to Bin.}

\item{A}{character vector of length 1 used to name the Bin. This name is used
to match column names when predicting on a \code{data.frame}.}
}
\value{
A single Bin objecte or a \link{\code{Classing}} object depending on
the value of \code{x} that was passed in.
}
\description{
Bin is the workhorse function of the Binnr2 package. At the
heart of the function is a fast discretization algorithm written in C that
optimizes information value given a set of constraints.
}

