% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,Classing-method}
\alias{[[}
\alias{[[,Classing-method}
\title{extract parts of a Classing}
\usage{
\S4method{[}{Classing,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{Classing,ANY,missing}(x, i, j, ...)
}
\description{
extract parts of a Classing

extract element of a Classing
}

